+++ tests/fixtures/48-zero-address-only/input.py
@@ -1,0 +1,3 @@
+from brownie import ZERO_ADDRESS
+
+NULL = ZERO_ADDRESS