+++ tests/fixtures/51-try-except-import/expected.py
@@ -1,0 +1,10 @@
+try:
+    from ape import accounts, networks
+except ImportError:
+    accounts = None
+    network = None
+
+
+def show_env():
+    if network is not None:
+        print(networks.active_provider.network.name)